fix:工时比例改为小数

+++ mes-common/common-model/src/main/java/com/richfit/mes/common/model/sys/TenantUser.java
@@ -85,7 +85,7 @@ public class TenantUser extends BaseEntity<TenantUser> {
     /**
      * 工时比例
      */
-    private Integer ratioHours;
+    private Double ratioHours;
 
 
     @TableField(exist = false)
